<FEAT> Deixando interceptor global.

+++ src/common/services/formatHal.service.ts
@@ -3,7 +3,7 @@ import { Injectable } from "@nestjs/common";
 @Injectable()
 export class FormatHalService {
     formatHal(data: any, parametersHal: any): any {
-        if(data instanceof Array && data.length > 0) {
+        if (data instanceof Array && data.length > 0) {
             return {
                 ...parametersHal,
                 _embedded: {
@@ -25,23 +25,25 @@ export class FormatHalService {
                     })
                 }
             }
-        } 
-        return {
-            ...parametersHal,
-            _embedded: {
-                ...data,
-                _links: {
-                    self: { href: `${parametersHal.self.href}/${data.id}` },
-                    edit: {
-                        href: `${parametersHal.self.href}/${data.id}`,
-                        method: 'PUT'
-                    },
-                    delete: {
-                        href: `${parametersHal.self.href}/${data.id}`,
-                        method: 'DELETE'
+        } else if (data) {
+            return {
+                ...parametersHal,
+                _embedded: {
+                    ...data,
+                    _links: {
+                        self: { href: `${parametersHal.self.href}/${data.id}` },
+                        edit: {
+                            href: `${parametersHal.self.href}/${data.id}`,
+                            method: 'PUT'
+                        },
+                        delete: {
+                            href: `${parametersHal.self.href}/${data.id}`,
+                            method: 'DELETE'
+                        }
                     }
                 }
             }
         }
+        return null;
     }
 }
